refact: move all business logics to content container
