+++ src/components/ContentContainer.js
@@ -1,3 +1,7 @@
+import { StationManager } from "./station-manager/StationManager.js";
+import { LineManager } from "./line-manager/LineManager.js";
+import { MapPrint } from "./map-print-manager/MapPrint.js";
+import { SectionManager } from "./section-manager/SectionManager.js";
 import { displayShow, displayhide } from "../utils/handleDom.js";
 
 const contentIds = [
@@ -7,8 +11,31 @@ const contentIds = [
   "map-print-manager-container",
 ];
 export class ContentContainer {
-  constructor() {}
+  constructor(props) {
+    this.props = {
+      ...props,
+      setStations: this.setStations,
+      getStations: this.getStations,
+      deleteStation: this.deleteStation,
+      setNewLine: this.setNewLine,
+      setLines: this.setLines,
+      getLines: this.getLines,
+      deleteLine: this.deleteLine,
+    };
+    this.initiateData();
+    this.initiateDOM();
+  }
 
+  initiateDOM = () => {
+    this.stationManager = new StationManager(this.props);
+    this.lineManager = new LineManager(this.props);
+    this.mapPrintManager = new MapPrint(this.props);
+    this.sectionManager = new SectionManager(this.props);
+  };
+  initiateData = () => {
+    localStorage.setItem("stations", JSON.stringify([]) || []);
+    localStorage.setItem("lines", JSON.stringify([]) || []);
+  };
   render = ({ id }) => {
     contentIds.forEach((contentId) => {
       const content = document.getElementById(contentId);
@@ -18,5 +45,60 @@ export class ContentContainer {
         displayhide(content);
       }
     });
+  };
+
+  setStations = (names) => {
+    localStorage.setItem("stations", JSON.stringify(names));
+    this.updateStationView();
+  };
+
+  getStations = () => {
+    return JSON.parse(localStorage.getItem("stations"));
+  };
+
+  deleteStation = (target) => {
+    let stations = this.getStations();
+    stations = stations.filter((station) => {
+      return station !== target;
+    });
+    this.setStations(stations);
+  };
+
+  updateStationView = () => {
+    this.lineManager.updateStations();
+    this.sectionManager.updateStationsInInput();
+  };
+
+  setLines = (lines) => {
+    localStorage.setItem("lines", JSON.stringify(lines));
+    this.updateLineView();
+  };
+
+  setNewLine = (lineName, newLines) => {
+    let lines = this.getLines();
+    let newLine = { lineName: lineName, stations: newLines };
+
+    lines.push(newLine);
+    this.setLines(lines);
+  };
+
+  getLines = () => {
+    return JSON.parse(localStorage.getItem("lines"));
+  };
+
+  deleteLine = (lineName) => {
+    let lines = this.getLines();
+    let newLines = lines.filter((line) => {
+      return line.lineName != lineName;
+    });
+
+    this.setLines(newLines);
+  };
+
+  updateLineView = () => {
+    this.sectionManager.updateHeaderButtons();
+    this.sectionManager.updateSectionList();
+    this.lineManager.updateLines();
+    this.mapPrintManager.render();
   };
 }
