+++ src/components/SubwayMap.js
@@ -1,9 +1,5 @@
 import { HeaderButtons } from "./HeaderButtons.js";
 import { ContentContainer } from "./ContentContainer.js";
-import { StationManager } from "./station-manager/StationManager.js";
-import { LineManager } from "./line-manager/LineManager.js";
-import { MapPrint } from "./map-print-manager/MapPrint.js";
-import { SectionManager } from "./section-manager/SectionManager.js";
 
 const buttonContentMap = {
   "station-manager-button": "station-manager-container",
@@ -16,91 +12,26 @@ export class SubwayMap {
   constructor() {
     localStorage.clear(); //지우고제출
     this.props = {
-      setStations: this.setStations,
-      getStations: this.getStations,
-      deleteStation: this.deleteStation,
-      setNewLine: this.setNewLine,
-      setLines: this.setLines,
-      getLines: this.getLines,
-      deleteLine: this.deleteLine,
+      clickHeaders: this.onHeaderClick,
     };
-    this.initiateData();
     this.initiateDOM();
+    this.render(this.props);
   }
 
-  initiateData = () => {
-    localStorage.setItem("stations", JSON.stringify([]));
-    localStorage.setItem("lines", JSON.stringify([]));
+  initiateDOM = () => {
+    this.content = new ContentContainer();
+    this.header = new HeaderButtons();
   };
 
-  initiateDOM = () => {
-    this.contentContainer = new ContentContainer();
-    new HeaderButtons({ clickHeaders: this.onHeaderClick });
-    this.stationManager = new StationManager(this.props);
-    this.lineManager = new LineManager(this.props);
-    this.mapPrintManager = new MapPrint(this.props);
-    this.sectionManager = new SectionManager(this.props);
+  render = () => {
+    this.header.render(this.props);
+    this.content.render(this.props);
   };
 
   onHeaderClick = (e) => {
     const { id } = e.currentTarget;
     const contentId = buttonContentMap[id];
 
-    this.contentContainer.render({ id: contentId });
-  };
-
-  setStations = (names) => {
-    localStorage.setItem("stations", JSON.stringify(names));
-    this.updateStationView();
-  };
-
-  getStations = () => {
-    return JSON.parse(localStorage.getItem("stations"));
-  };
-
-  deleteStation = (target) => {
-    let stations = this.getStations();
-    stations = stations.filter((station) => {
-      return station !== target;
-    });
-    this.setStations(stations);
-  };
-
-  updateStationView = () => {
-    this.lineManager.updateStations();
-    this.sectionManager.updateStationsInInput();
-  };
-
-  setLines = (lines) => {
-    localStorage.setItem("lines", JSON.stringify(lines));
-    this.updateLineView();
-  };
-
-  setNewLine = (lineName, newLines) => {
-    let lines = this.getLines();
-    let newLine = { lineName: lineName, stations: newLines };
-
-    lines.push(newLine);
-    this.setLines(lines);
-  };
-
-  getLines = () => {
-    return JSON.parse(localStorage.getItem("lines"));
-  };
-
-  deleteLine = (lineName) => {
-    let lines = this.getLines();
-    let newLines = lines.filter((line) => {
-      return line.lineName != lineName;
-    });
-
-    this.setLines(newLines);
-  };
-
-  updateLineView = () => {
-    this.sectionManager.updateHeaderButtons();
-    this.sectionManager.updateSectionList();
-    this.lineManager.updateLines();
-    this.mapPrintManager.render();
+    this.content.render({ id: contentId });
   };
 }
